requirements.txt
streamlit
openai>=1.0.0
beautifulsoup4
requests
python-dotenv
